Move back jobs logic to worker.js
