--- lib/util/jobs.js
@@ -1,36 +1,0 @@
-const ms = require('ms')
-
-const {runWorkflow} = require('../worker/run-workflow')
-const {harvestRequestedSources} = require('../worker/start-harvest')
-
-const jobsList = [
-  {
-    name: 'lancement du workflow',
-    every: '1h',
-    async handler() {
-      await runWorkflow()
-    }
-  },
-  {
-    name: 'recherche sources à moissonner',
-    every: '30s',
-    async handler() {
-      await harvestRequestedSources()
-    }
-  }
-]
-
-function startJobs(jobs) {
-  jobs.forEach(job => {
-    setInterval(() => {
-      const now = new Date()
-      console.log(`${now.toISOString().slice(0, 19)} | running job : ${job.name}`)
-      job.handler()
-    }, ms(job.every))
-  })
-}
-
-module.exports = {
-  jobsList,
-  startJobs
-}
