+++ worker.js
@@ -1,15 +1,41 @@
 #!/usr/bin/env node --max_old_space_size=8192
 require('dotenv').config()
 
+const ms = require('ms')
+
 const mongo = require('./lib/util/mongo')
 const {endFarms} = require('./lib/util/farms')
 const {runWorkflow} = require('./lib/worker/run-workflow')
-const {jobsList, startJobs} = require('./lib/util/jobs')
+const {harvestRequestedSources} = require('./lib/worker/start-harvest')
+
+const jobsList = [
+  {
+    name: 'lancement du workflow',
+    every: '1h',
+    async handler() {
+      await runWorkflow()
+    }
+  },
+  {
+    name: 'recherche sources à moissonner',
+    every: '30s',
+    async handler() {
+      await harvestRequestedSources()
+    }
+  }
+]
 
 async function main() {
   await mongo.connect()
   await runWorkflow()
-  startJobs(jobsList)
+
+  jobsList.forEach(jobType => {
+    setInterval(() => {
+      const now = new Date()
+      console.log(`${now.toISOString().slice(0, 19)} | running job : ${jobType.name}`)
+      jobType.handler()
+    }, ms(jobType.every))
+  })
 }
 
 main().catch(error => {
